Encode Clusters, ClusterQCs and PubKeys when deploying Epochs contract
